server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib
bcrypt
python-dotenv
google-generativeai
python-multipart
